+++ libs/ledgerjs/packages/types-live/src/feature.ts
@@ -178,6 +178,7 @@ export type Features = CurrencyFeatures & {
   cexDepositEntryPointsMobile: Feature_CexDepositEntryPointsMobile;
   fetchAdditionalCoins: Feature_FetchAdditionalCoins;
   ptxSwapLiveApp: Feature_PtxSwapLiveApp;
+  ptxSwapMoonpayProvider: Feature_PtxSwapMoonpayProvider;
 };
 
 /**
@@ -489,6 +490,7 @@ export type Feature_ListAppsV2minor1 = DefaultFeature;
 export type Feature_BrazeLearn = DefaultFeature;
 export type Feature_LlmNewDeviceSelection = DefaultFeature;
 export type Feature_LlmWalletQuickActions = DefaultFeature;
+export type Feature_PtxSwapMoonpayProvider = DefaultFeature;
 
 /**
  * Utils types.